refactor: extract parseAndValidate helper to remove backend duplication
The same try/catch/JSON-parse/safeParse block was copied in all four
backends. Extracted to src/core/parse.ts with an extractJson option
for backends (Anthropic) that may receive prose-wrapped JSON responses.
