+++ src/backends/anthropic.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 import type { ShapecraftModel } from "../types.js";
-import { SchemaViolationError } from "../types.js";
 import { buildStructuredPrompt } from "../core/schema.js";
+import { parseAndValidate } from "../core/parse.js";
 
 export interface AnthropicBackendOptions {
   model?: string;
@@ -40,15 +40,7 @@ export function anthropic(options: AnthropicBackendOptions = {}): ShapecraftMode
       const raw: string =
         response.content[0]?.type === "text" ? response.content[0].text : "";
 
-      try {
-        const jsonMatch = raw.match(/\{[\s\S]*\}|\[[\s\S]*\]/);
-        const parsed = JSON.parse(jsonMatch?.[0] ?? raw);
-        const result = schema.safeParse(parsed);
-        if (!result.success) throw new SchemaViolationError(raw, result.error);
-        return result.data as T;
-      } catch (err) {
-        throw new SchemaViolationError(raw, err);
-      }
+      return parseAndValidate<T>(raw, schema, { extractJson: true });
     },
   };
 }
