+++ src/backends/ollama.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 import type { ShapecraftModel } from "../types.js";
-import { SchemaViolationError } from "../types.js";
 import { toJsonSchema, buildStructuredPrompt } from "../core/schema.js";
+import { parseAndValidate } from "../core/parse.js";
 
 export interface OllamaBackendOptions {
   model: string;
@@ -40,14 +40,7 @@ export function ollama(options: OllamaBackendOptions): ShapecraftModel {
       const json = await response.json() as { message?: { content?: string } };
       const raw = json.message?.content ?? "";
 
-      try {
-        const parsed = JSON.parse(raw);
-        const result = schema.safeParse(parsed);
-        if (!result.success) throw new SchemaViolationError(raw, result.error);
-        return result.data as T;
-      } catch (err) {
-        throw new SchemaViolationError(raw, err);
-      }
+      return parseAndValidate<T>(raw, schema);
     },
   };
 }
